pyproject.toml
[project]
name = "utah-watchdog"
version = "0.1.0"
description = "Track Utah legislators - votes, money, legislation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.2",
    "httpx>=0.28.1",
    "motor>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.6.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
